requirements.txt
aiocontextvars==0.2.2
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
contextvars==2.4
Django==3.0.8
Flask==1.1.2
idna==2.10
image==1.5.32
immutables==0.14
itsdangerous==1.1.0
Jinja2==2.11.2
loguru==0.5.1
MarkupSafe==1.1.1
numpy==1.19.0
Pillow==7.2.0
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
Werkzeug==1.0.1
